Made manipulate handler callbacks consistent.
When {pre,post}Manipulate was specified as an array, the callback
expected only the content, but when it was just a single handler
it expected an error as well as the content.

I would have preferred to use the nodejs pattern and include the error
but all connect-assetmanager-handlers only pass content so I went the
simpler route for now.